Project Title: FIFA World Cup
Project Description:
For the FIFA World Cup 2006, a fly-over has been constructed between the hotel
where the teams are staying and the stadium. This fly-over will be used by the
German team and the Italian team in the upcoming semifinal on Tuesday. A tram
car is used to cross this fly-over, but it seats only four people, and must always
carry a full load. We cannot put three Italians and one German in the same tramcar,
because the Italians would be in majority and might try to intimidate the German.
Similarly, we cannot put three Germans in the same tram-car with one Italian. All
other combinations are safe. Implement a synchronization method to solve the
problem. You must output the progress of the processes. E.g., When a player
arrives you should output who arrived and also print the total number of Germans
and Italians waiting to board the tram. When a tram leaves you should output that
as well.